+++ ChangeLog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved handling corrupted glTF files (thanks [@zharry][zharry] for #230)
 - Loading [Ready Player Me][ReadyPlayerMe] avatars with unsupported node extension (`MOZ_hubs_components`) 
 - Loading glTF-binary files that have no buffers or an empty binary chunk (#227)
+- Crash and incorrect mesh clustering caused in `MeshPrimitive.Equals` (#224)
 
 ## [4.2.0] - 2021-07-16
 ### Added